Moved jest globals in single folder.
