+++ src/engine/core/managers/settings/GameSettingsManager.test.ts
@@ -4,7 +4,7 @@ import { level } from "xray16";
 import { disposeManager, getManagerInstance, registry } from "@/engine/core/database";
 import { GameSettingsManager } from "@/engine/core/managers/settings/GameSettingsManager";
 import { EGameDifficulty, gameDifficulties } from "@/engine/lib/constants/game_difficulties";
-import { replaceFunctionMock } from "@/fixtures/utils/function_mock";
+import { replaceFunctionMock } from "@/fixtures/jest";
 import { gameConsole } from "@/fixtures/xray/mocks/console.mock";
 import { EPacketDataType, mockNetPacket, mockNetProcessor, MockNetProcessor } from "@/fixtures/xray/mocks/save";
