+++ src/engine/core/utils/game/game_time.test.ts
@@ -11,7 +11,7 @@ import {
 } from "@/engine/core/utils/game/game_time";
 import { MAX_I32, MAX_U8, MIN_I32 } from "@/engine/lib/constants/memory";
 import { Optional, Time } from "@/engine/lib/types";
-import { replaceFunctionMock } from "@/fixtures/utils/function_mock";
+import { replaceFunctionMock } from "@/fixtures/jest";
 import { MockCTime } from "@/fixtures/xray/mocks/CTime.mock";
 import { EPacketDataType, mockNetPacket, mockNetProcessor, MockNetProcessor } from "@/fixtures/xray/mocks/save";
