+++ src/engine/core/utils/logging/LuaLogger.test.ts
@@ -4,7 +4,7 @@ import { log, print_stack, time_global } from "xray16";
 import { LuaLogger } from "@/engine/core/utils/logging/LuaLogger";
 import { toJSON } from "@/engine/core/utils/transform/json";
 import { gameConfig } from "@/engine/lib/configs/GameConfig";
-import { replaceFunctionMock, resetFunctionMock } from "@/fixtures/utils";
+import { replaceFunctionMock, resetFunctionMock } from "@/fixtures/jest";
 
 describe("LuaLogger class", () => {
   const logging: { LuaLogger: typeof LuaLogger } = jest.requireActual("@/engine/core/utils/logging/LuaLogger");
