+++ src/engine/core/utils/object/object_state.test.ts
@@ -10,7 +10,7 @@ import { WoundManager } from "@/engine/core/schemes/stalker/wounded/WoundManager
 import { isObjectAsleep, isObjectMeeting, isObjectWounded } from "@/engine/core/utils/object/object_state";
 import { ActionPlanner, ClientObject, EScheme } from "@/engine/lib/types";
 import { mockSchemeState } from "@/fixtures/engine";
-import { replaceFunctionMock } from "@/fixtures/utils";
+import { replaceFunctionMock } from "@/fixtures/jest";
 import { mockClientGameObject } from "@/fixtures/xray";
 
 describe("object state utils", () => {
