+++ src/engine/scripts/register/class_registrator.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, jest } from "@jest/globals";
 
 import { EClientObjectClass } from "@/engine/lib/constants/class_ids";
 import { registerGameClasses } from "@/engine/scripts/register/class_registrator";
-import { getFunctionMock } from "@/fixtures/utils";
+import { getFunctionMock } from "@/fixtures/jest";
 import { mockObjectFactory } from "@/fixtures/xray/mocks/objects/ObjectFactory.mock";
 
 describe("'class_registrator' entry point", () => {
