+++ src/fixtures/jest/index.ts
@@ -1,1 +1,2 @@
-export * from "@/fixtures/jest/extendJest";
+export * from "@/fixtures/jest/utils";
+export * from "@/fixtures/jest/extend_jest";
