+++ src/fixtures/jest/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "@/fixtures/jest/utils/function_mock";
