--- src/fixtures/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "@/fixtures/utils/function_mock";
